+++ CHANGELOG.md
@@ -26,6 +26,8 @@ This file is used to list changes made in each version of the AWS ParallelCluste
 - Restore ``noatime`` option, which has positive impact on the performances of NFS filesystem
 - Upgrade NICE DCV to version 2021.1-10598.
 - Upgrade Slurm to version 20.11.8
+- Upgrade NVIDIA driver to version 470.57.02.
+- Upgrade CUDA library to version 11.4.0.
 
 2.11.0
 -----